+++ examples/react/src/pages/index.tsx
@@ -1,3 +1,14 @@
+/**
+ * @handle {
+ * "title": "首页",
+ * "requiresAuth": false,
+ * "const": true,
+ * "meta": {
+ *   "description": "主页"
+ *   }
+ * }
+ */
+
 import { Navigate } from 'react-router-dom'
 
 function Index() {