+++ examples/react/src/pages/test/index.tsx
@@ -1,0 +1,16 @@
+/**
+ * @handle {
+ *  "title": "首页",
+ *  "requiresAuth": false,
+ *  "const": true,
+ *  "meta": {
+ *    "description": "我是家的感觉你懂！！！！"
+ *  }
+ * }
+ */
+
+export default function Index() {
+  return (
+    <div>888888</div>
+  )
+}