+++ examples/react/src/pages/test/layout.tsx
@@ -1,0 +1,10 @@
+import { Outlet } from 'react-router'
+
+export default function RootLayout() {
+  return (
+    <div style={{ background: 'linear-gradient(135deg,rgb(234, 175, 102) 0%, #764ba2 100%)' }}>
+      66666
+      <Outlet />
+    </div>
+  )
+}